% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHDImatrix.R
\name{get_DH}
\alias{get_DH}
\title{Get the demographic history from each genealogy}
\usage{
get_DH(i, tree, lower_HDI, epsilon = 0, n = 1000)
}
\arguments{
\item{i}{which number of the phylogenetic tree need to be analyzed.}

\item{tree}{either an ultra-metric phylogenetic tree (i.e. an object of class "phylo") or a vector of interval lengths.}

\item{lower_HDI}{Lower HDI value of the phylogenetic tree.}

\item{epsilon}{collapsing parameter that controls the amount of smoothing.}

\item{n}{A number of intervals used to build the matrix.}
}
\value{
Normalized coalescent interval and effective population size for the phylogenetic tree.
}
\description{
This function was used to obtain the demographic history from the genealogy
Uses classic skyline implemented in ape package to find the time at the end of each coalescent interval and effective population size of each interval.
}
