% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHDImatrix.R
\name{get_HDI_matrix}
\alias{get_HDI_matrix}
\title{Get a HDI matrix for the phylogenetic tree}
\usage{
get_HDI_matrix(dat, margin = 2, ci = 0.95)
}
\arguments{
\item{dat}{Data matrix used for extracting highest density intervals, and assume that each row represents a sample and each column show the coalescent interval.}

\item{margin}{Default by column.}

\item{ci}{The value of credible interval.}
}
\value{
A HDI matrix with each column is the value of lower, median, and upper bound.
}
\description{
This function was used to obtain a HDI matrix of the phylogenetic tree.
Uses hdi (Highest Density Interval) algorithm as implemented in HDInterval package.
}
