% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHDImatrix.R
\name{get_lower_HDI}
\alias{get_lower_HDI}
\title{Get Lower HDI value of the phylogenetic tree}
\usage{
get_lower_HDI(tree, ci = 0.95)
}
\arguments{
\item{tree}{either an ultra-metric phylogenetic tree (i.e. an object of class "phylo") or a vector of interval lengths.}

\item{ci}{The value of credible interval.}
}
\value{
lower limits of the HDI for all trees in the tree file.
}
\description{
This function was used to calculate the lower HDI value of the phylogenetic tree.
Uses hdi (Highest Density Interval) algorithm as implemented in HDInterval package.
Uses coalescent.intervals as implemented in ape package.
}
