% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHDImatrix.R
\name{get_slide_matrix}
\alias{get_slide_matrix}
\title{Get a smooth HDI matrix by sliding window method}
\usage{
get_slide_matrix(x, w = 100, s = 20)
}
\arguments{
\item{x}{Data matrix of each row represents a sample and each column show the coalescent interval.}

\item{w}{window size of the sliding window.}

\item{s}{step size of the sliding window.}
}
\value{
A smooth HDI matrix with each column is the value of lower, median, and upper bound.
}
\description{
This function was used to obtain a smooth HDI matrix by sliding window.
Uses zoo algorithm as implemented in zoo package.
}
