% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pair_data}
\alias{pair_data}
\title{Distance data matrix for 16 samples including the heteroduplex mobility data and nucleotide sequencing data.}
\format{
A data frame with 120 rows and 2 variables:
\describe{
   \item{ober}{Heteroduplex mobility data}
   \item{real}{Nucleotide sequencing data}
   ...
}
}
\usage{
pair_data
}
\description{
Pairwise distance data matrix for 16 samples
}
\keyword{datasets}
