% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_bootstrap.R
\name{para_boot}
\alias{para_boot}
\title{Parametric bootstrap sampling using the predict interval}
\usage{
para_boot(pair_data, dframe, method)
}
\arguments{
\item{pair_data}{Assume by default that the data matrix is the pairwise distances matrix, which contain two column named ober (HMA data) and real (DNA data).}

\item{dframe}{A data frame for the samples.}

\item{method}{only allow "nj" or "upgma" method to build the phylogeny.}
}
\value{
the phylogenetic tree of the parametric bootstrap.
}
\description{
This function was used to bootstrap sampling for HMA data using the predict interval.
Uses neighbour-joining (nj) algorithm as implemented in ape package.
Uses upgma method as implemented in phangorn package.
}
