% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lm.R
\name{plot_lm}
\alias{plot_lm}
\title{Plot the linear regression with prediction interval}
\usage{
plot_lm(pair_data)
}
\arguments{
\item{pair_data}{Assume by default that the data matrix is the pairwise distances matrix, which contain two column named ober (HMA data) and real (DNA data).}
}
\description{
This function was used to perform the correlation analysis between HMA and nucleotide sequencing data
Uses ggplot2 package.
}
\examples{
data(pair_data)
plot_lm(pair_data)

}
